Split fill tests to header and source
Seperate header and source to enable datapar tests.
