+++ libs/parallelism/algorithms/tests/unit/algorithms/fill.cpp
@@ -1,92 +1,20 @@
 //  Copyright (c) 2014 Grant Mercer
 //  Copyright (c) 2017-2020 Hartmut Kaiser
+//  Copyright (c) 2021 Srinivas Yadav
 //
 //  SPDX-License-Identifier: BSL-1.0
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
 #include <hpx/local/init.hpp>
-#include <hpx/modules/testing.hpp>
-#include <hpx/parallel/algorithms/fill.hpp>
 
-#include <cstddef>
 #include <iostream>
-#include <iterator>
-#include <numeric>
-#include <random>
 #include <string>
 #include <vector>
 
-#include "test_utils.hpp"
+#include "fill_tests.hpp"
 
 ////////////////////////////////////////////////////////////////////////////
-unsigned int seed = std::random_device{}();
-std::mt19937 gen(seed);
-
-template <typename IteratorTag>
-void test_fill(IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c(10007);
-    std::iota(std::begin(c), std::end(c), gen());
-
-    hpx::fill(iterator(std::begin(c)), iterator(std::end(c)), 10);
-
-    // verify values
-    std::size_t count = 0;
-    std::for_each(std::begin(c), std::end(c), [&count](std::size_t v) -> void {
-        HPX_TEST_EQ(v, std::size_t(10));
-        ++count;
-    });
-    HPX_TEST_EQ(count, c.size());
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_fill(ExPolicy policy, IteratorTag)
-{
-    static_assert(hpx::is_execution_policy<ExPolicy>::value,
-        "hpx::is_execution_policy<ExPolicy>::value");
-
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c(10007);
-    std::iota(std::begin(c), std::end(c), gen());
-
-    hpx::fill(policy, iterator(std::begin(c)), iterator(std::end(c)), 10);
-
-    // verify values
-    std::size_t count = 0;
-    std::for_each(std::begin(c), std::end(c), [&count](std::size_t v) -> void {
-        HPX_TEST_EQ(v, std::size_t(10));
-        ++count;
-    });
-    HPX_TEST_EQ(count, c.size());
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_fill_async(ExPolicy p, IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c(10007);
-    std::iota(std::begin(c), std::end(c), gen());
-
-    hpx::future<void> f =
-        hpx::fill(p, iterator(std::begin(c)), iterator(std::end(c)), 10);
-    f.wait();
-
-    std::size_t count = 0;
-    std::for_each(std::begin(c), std::end(c), [&count](std::size_t v) -> void {
-        HPX_TEST_EQ(v, std::size_t(10));
-        ++count;
-    });
-    HPX_TEST_EQ(count, c.size());
-}
-
 template <typename IteratorTag>
 void test_fill()
 {
@@ -108,109 +36,7 @@ void fill_test()
     test_fill<std::forward_iterator_tag>();
 }
 
-///////////////////////////////////////////////////////////////////////////////
-template <typename IteratorTag>
-void test_fill_exception(IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::decorated_iterator<base_iterator, IteratorTag>
-        decorated_iterator;
-    std::vector<std::size_t> c(10007);
-    std::iota(std::begin(c), std::end(c), gen());
-
-    bool caught_exception = false;
-    try
-    {
-        hpx::fill(decorated_iterator(std::begin(c),
-                      []() { throw std::runtime_error("test"); }),
-            decorated_iterator(std::end(c)), 10);
-        HPX_TEST(false);
-    }
-    catch (hpx::exception_list const& e)
-    {
-        caught_exception = true;
-        test::test_num_exceptions<hpx::execution::sequenced_policy,
-            IteratorTag>::call(hpx::execution::seq, e);
-    }
-    catch (...)
-    {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_exception);
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_fill_exception(ExPolicy policy, IteratorTag)
-{
-    static_assert(hpx::is_execution_policy<ExPolicy>::value,
-        "hpx::is_execution_policy<ExPolicy>::value");
-
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::decorated_iterator<base_iterator, IteratorTag>
-        decorated_iterator;
-    std::vector<std::size_t> c(10007);
-    std::iota(std::begin(c), std::end(c), gen());
-
-    bool caught_exception = false;
-    try
-    {
-        hpx::fill(policy,
-            decorated_iterator(
-                std::begin(c), []() { throw std::runtime_error("test"); }),
-            decorated_iterator(std::end(c)), 10);
-        HPX_TEST(false);
-    }
-    catch (hpx::exception_list const& e)
-    {
-        caught_exception = true;
-        test::test_num_exceptions<ExPolicy, IteratorTag>::call(policy, e);
-    }
-    catch (...)
-    {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_exception);
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_fill_exception_async(ExPolicy p, IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::decorated_iterator<base_iterator, IteratorTag>
-        decorated_iterator;
-
-    std::vector<std::size_t> c(10007);
-    std::iota(std::begin(c), std::end(c), gen());
-
-    bool caught_exception = false;
-    bool returned_from_algorithm = false;
-    try
-    {
-        hpx::future<void> f = hpx::fill(p,
-            decorated_iterator(
-                std::begin(c), []() { throw std::runtime_error("test"); }),
-            decorated_iterator(std::end(c)), 10);
-        returned_from_algorithm = true;
-        f.get();
-
-        HPX_TEST(false);
-    }
-    catch (hpx::exception_list const& e)
-    {
-        caught_exception = true;
-        test::test_num_exceptions<ExPolicy, IteratorTag>::call(p, e);
-    }
-    catch (...)
-    {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_exception);
-    HPX_TEST(returned_from_algorithm);
-}
-
+////////////////////////////////////////////////////////////////////////////
 template <typename IteratorTag>
 void test_fill_exception()
 {
@@ -234,75 +60,7 @@ void fill_exception_test()
     test_fill_exception<std::forward_iterator_tag>();
 }
 
-//////////////////////////////////////////////////////////////////////////////
-template <typename ExPolicy, typename IteratorTag>
-void test_fill_bad_alloc(ExPolicy policy, IteratorTag)
-{
-    static_assert(hpx::is_execution_policy<ExPolicy>::value,
-        "hpx::is_execution_policy<ExPolicy>::value");
-
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::decorated_iterator<base_iterator, IteratorTag>
-        decorated_iterator;
-
-    std::vector<std::size_t> c(100007);
-    std::iota(std::begin(c), std::end(c), gen());
-
-    bool caught_bad_alloc = false;
-    try
-    {
-        hpx::fill(policy,
-            decorated_iterator(std::begin(c), []() { throw std::bad_alloc(); }),
-            decorated_iterator(std::end(c)), 10);
-        HPX_TEST(false);
-    }
-    catch (std::bad_alloc const&)
-    {
-        caught_bad_alloc = true;
-    }
-    catch (...)
-    {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_bad_alloc);
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_fill_bad_alloc_async(ExPolicy p, IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::decorated_iterator<base_iterator, IteratorTag>
-        decorated_iterator;
-
-    std::vector<std::size_t> c(10007);
-    std::iota(std::begin(c), std::end(c), gen());
-
-    bool caught_bad_alloc = false;
-    bool returned_from_algorithm = false;
-    try
-    {
-        hpx::future<void> f = hpx::fill(p,
-            decorated_iterator(std::begin(c), []() { throw std::bad_alloc(); }),
-            decorated_iterator(std::end(c)), 10);
-        returned_from_algorithm = true;
-        f.get();
-
-        HPX_TEST(false);
-    }
-    catch (std::bad_alloc const&)
-    {
-        caught_bad_alloc = true;
-    }
-    catch (...)
-    {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_bad_alloc);
-    HPX_TEST(returned_from_algorithm);
-}
-
+////////////////////////////////////////////////////////////////////////////
 template <typename IteratorTag>
 void test_fill_bad_alloc()
 {
